pyproject.toml
[project]
name = "refractiveindex-info-database"
version = "2025.02.23"
description = "Database of optical constants"
requires-python = ">=3.10"
dependencies = [
    "cerberus>=1.3.7",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "pyqt6>=6.8.1",
    "pytest>=8.3.5",
    "pytest-xdist>=3.6.1",
    "pyyaml>=6.0.2",
]

[tool.pytest.ini_options]
addopts = "-n 12"
